+++ igata/settings.py
@@ -2,7 +2,6 @@
 Project/package-wide settings
 """
 import datetime
-import json
 import logging
 import os
 
@@ -70,6 +69,9 @@ INPUT_CSV_ENCODING = os.getenv("INPUT_CSV_ENCODING", DEFAULT_INPUT_CSV_ENCODING)
 
 DEFAULT_INPUT_CSV_DELIMITER = ","
 INPUT_CSV_DELIMITER = os.getenv("INPUT_CSV_DELIMITER", DEFAULT_INPUT_CSV_DELIMITER)
+
+DEFAULT_OUTPUT_CSV_ENCODING = os.getenv("DEFAULT_OUTPUT_CSV_ENCODING", "utf8")
+DEFAULT_OUTPUT_CSV_DELIMITER = os.getenv("DEFAULT_OUTPUT_CSV_DELIMITER", ",")
 
 DEFAULT_INPUT_CSV_HEADER_LINES = "0"  # a header may span more than 1 line
 INPUT_CSV_HEADER_LINES = int(os.getenv("INPUT_CSV_HEADER_LINES", DEFAULT_INPUT_CSV_HEADER_LINES))
